Removed unused imports and converted timeout duration into double
Removed unused imports from each file.

Converted window setting time fields into doubles instead of ints to operate on a smaller, more precise scale similar to that of gaze metrics.

Limited event window event selections to a fixed amount and filtered by fixations whenever necessary for baseline values.

+++ src/analysis/WindowSettings.java
@@ -3,18 +3,18 @@ package analysis;
 public class WindowSettings {
 
     public boolean tumblingEnabled;
-    public int tumblingWindowSize;
+    public double tumblingWindowSize;
 
     public boolean expandingEnabled;
-    public int expandingWindowSize;
+    public double expandingWindowSize;
 
     public boolean hoppingEnabled;
-    public int hoppingWindowSize;
-    public int hoppingHopSize;
+    public double hoppingWindowSize;
+    public double hoppingHopSize;
     
     public boolean eventEnabled;
     public String event;
-    public int eventTimeout;
+    public double eventTimeout;
 
     public WindowSettings() {
         this.tumblingEnabled = false;
